% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluationMetrics.R
\name{MSE}
\alias{MSE}
\title{Calculate Mean Squared Error (MSE) score}
\usage{
MSE(model, decimal = 3)
}
\arguments{
\item{model}{The fitted linear model object}
}
\value{
MSE score in a numeric value with optional decimal places
}
\description{
Calculate MSE score by parsing the fitted linear model object
}
\examples{
MSE(fit.lm, decimal=3)
}
