% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluationMetrics.R
\name{RMSE}
\alias{RMSE}
\title{Calculate Root Mean Squared Error (RMSE) score}
\usage{
RMSE(model, decimal = 3)
}
\arguments{
\item{model}{The fitted linear model object}
}
\value{
RMSE score in a numeric value with optional decimal places
}
\description{
Calculate RMSE score by parsing the fitted linear model object
}
\examples{
RMSE(fit.lm, decimal=3)
}
