% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluationMetrics.R
\name{RSQ}
\alias{RSQ}
\title{Calculate R-squared score}
\usage{
RSQ(model, decimal = 3)
}
\arguments{
\item{model}{The fitted linear model object}
}
\value{
R-squared score in a numeric value with optional decimal places
}
\description{
Calculate R-squared score by parsing the fitted linear model object
}
\examples{
RSQ(fit.lm, decimal=3)
}
